requirements.txt
torch-ema
torchmetrics
ninja
trimesh
opencv-python
tensorboardX
numpy
pandas
tqdm
matplotlib
PyMCubes
rich
pysdf
dearpygui
packaging
scipy
lpips
imageio==2.13.0
torchmetrics
imageio-ffmpeg==0.4.8
open3d
configargparse
scikit-image
nksr
black

# nuscenes
nuscenes-devkit>=1.1.1
pyquaternion
